Moved WebSocket configuration
